refactor: TransactionTemplate으로 관리 중복 해결
기존에는 트랜잭션 관리를 서비스 계층에서 수행했다.

관리를 위한 중복 코드를 템플릿화한다. 또한 덕분에 서비스 계층에서는 DB
종속적인 예외를 다루지 않아도 된다.
