+++ src/main/java/controller/JanggiController.java
@@ -13,7 +13,6 @@ import dto.GameWrapper;
 import dto.LoadCommand;
 import dto.LoadCommand.Type;
 import dto.MoveCommand;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
