+++ src/main/java/service/TransactionCallback.java
@@ -1,0 +1,9 @@
+package service;
+
+import java.sql.Connection;
+
+@FunctionalInterface
+public interface TransactionCallback<T> {
+
+    T doInTransaction(Connection connection);
+}
