+++ src/main/java/service/TransactionTemplate.java
@@ -1,0 +1,52 @@
+package service;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class TransactionTemplate {
+
+    private final DataSource dataSource;
+
+    public TransactionTemplate(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public <T> T execute(TransactionCallback<T> action) {
+        Connection conn = null;
+        try {
+            conn = dataSource.getConnection();
+            conn.setAutoCommit(false);
+
+            T doingResult = action.doInTransaction(conn);
+
+            conn.commit();
+
+            return doingResult;
+        } catch (SQLException e) {
+            rollback(conn);
+            // TODO: 커스텀 런타임으로 변경
+            throw new RuntimeException(e);
+        } finally {
+            close(conn);
+        }
+    }
+
+    private static void rollback(Connection conn) {
+        try {
+            conn.rollback();
+        } catch (SQLException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    private static void close(Connection conn) {
+        if (conn != null) {
+            try {
+                conn.close();
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
+        }
+    }
+}
